refactor: remove obsolete components and configuration files from the project
