--- apps/oberyn/.vscode/settings.json
@@ -1,11 +1,0 @@
-{
-  "typescript.tsdk": "node_modules/typescript/lib",
-  "cSpell.words": [
-    "Dorne",
-    "Lançassolar",
-    "martell",
-    "messageformat",
-    "nextui",
-    "Nymeros"
-  ]
-}
