+++ apps/oberyn/src/components/MovieHero.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import { MovieHeroProps } from '@/types/types'
+import { MovieHeroProps } from '@/src/types/types'
 
 const MovieHero: React.FC<MovieHeroProps> = ({
   Title,
