+++ apps/oberyn/src/public/favicon.ico
